regionQuery

+++ DBscan/src/br/com/ufc/quixada/dbscan/model/Point.java
@@ -1,6 +1,7 @@
 package br.com.ufc.quixada.dbscan.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public class Point implements Serializable{
@@ -9,8 +10,8 @@ public class Point implements Serializable{
 	private Integer idPonto;
 	private Integer idTaxi;
 	private Integer idCluster;
-	private Long longitude;
-	private Long latitude;
+	private BigDecimal longitude;
+	private BigDecimal latitude;
 	private boolean isNoise;
 	private boolean isCore;
 	private Timestamp time;
@@ -22,64 +23,83 @@ public class Point implements Serializable{
 		this.visit = false;
 		this.dist = 0.0;
 	}
-	
+
 	public Integer getIdPonto() {
 		return idPonto;
 	}
+
 	public void setIdPonto(Integer idPonto) {
 		this.idPonto = idPonto;
 	}
+
 	public Integer getIdTaxi() {
 		return idTaxi;
 	}
+
 	public void setIdTaxi(Integer idTaxi) {
 		this.idTaxi = idTaxi;
 	}
+
 	public Integer getIdCluster() {
 		return idCluster;
 	}
+
 	public void setIdCluster(Integer idCluster) {
 		this.idCluster = idCluster;
 	}
-	public Long getLongitude() {
+
+	public BigDecimal getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(Long longitude) {
+
+	public void setLongitude(BigDecimal longitude) {
 		this.longitude = longitude;
 	}
-	public Long getLatitude() {
+
+	public BigDecimal getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(Long latitude) {
+
+	public void setLatitude(BigDecimal latitude) {
 		this.latitude = latitude;
 	}
+
 	public boolean isNoise() {
 		return isNoise;
 	}
+
 	public void setNoise(boolean isNoise) {
 		this.isNoise = isNoise;
 	}
+
 	public boolean isCore() {
 		return isCore;
 	}
+
 	public void setCore(boolean isCore) {
 		this.isCore = isCore;
 	}
+
 	public Timestamp getTime() {
 		return time;
 	}
+
 	public void setTime(Timestamp time) {
 		this.time = time;
 	}
+
 	public Double getDist() {
 		return dist;
 	}
+
 	public void setDist(Double dist) {
 		this.dist = dist;
 	}
+
 	public boolean isVisit() {
 		return visit;
 	}
+
 	public void setVisit(boolean visit) {
 		this.visit = visit;
 	}
@@ -90,6 +110,5 @@ public class Point implements Serializable{
 				+ longitude + ", latitude=" + latitude + ", isNoise=" + isNoise + ", isCore=" + isCore + ", time="
 				+ time + ", dist=" + dist + ", visit=" + visit + "]";
 	}
-	
 	
 }
